[inner] decision_tree
