+++ algorithms/kernel/decision_tree/decision_tree_classification_train_dense_default_batch_fpt_cpu.cpp
@@ -34,10 +34,6 @@ namespace classification
 {
 namespace training
 {
-namespace interface1
-{
-template class BatchContainer<DAAL_FPTYPE, defaultDense, DAAL_CPU>;
-} // namespace interface1
 
 namespace interface2
 {
@@ -47,7 +43,6 @@ template class BatchContainer<DAAL_FPTYPE, defaultDense, DAAL_CPU>;
 namespace internal
 {
 
-template class DecisionTreeTrainBatchKernel<DAAL_FPTYPE, decision_tree::classification::interface1::Parameter, defaultDense, DAAL_CPU>;
 template class DecisionTreeTrainBatchKernel<DAAL_FPTYPE, decision_tree::classification::Parameter, defaultDense, DAAL_CPU>;
 
 } // namespace internal
