+++ algorithms/kernel/decision_tree/decision_tree_regression_train_dense_default_batch_fpt_cpu.cpp
@@ -34,11 +34,6 @@ namespace regression
 {
 namespace training
 {
-namespace interface1
-{
-template class BatchContainer<DAAL_FPTYPE, defaultDense, DAAL_CPU>;
-} // namespace interface1
-
 namespace interface2
 {
 template class BatchContainer<DAAL_FPTYPE, defaultDense, DAAL_CPU>;
@@ -46,9 +41,7 @@ template class BatchContainer<DAAL_FPTYPE, defaultDense, DAAL_CPU>;
 
 namespace internal
 {
-
 template class DecisionTreeTrainBatchKernel<DAAL_FPTYPE, defaultDense, DAAL_CPU>;
-
 } // namespace internal
 } // namespace training
 } // namespace regression
