+++ algorithms/kernel/decision_tree/inner/decision_tree_classification_predict_dense_default_batch_container_v1.h
@@ -1,4 +1,4 @@
-/* file: decision_tree_classification_predict_dense_default_batch_container.h */
+/* file: decision_tree_classification_predict_dense_default_batch_container_v1.h */
 /*******************************************************************************
 * Copyright 2014-2019 Intel Corporation
 *
@@ -35,7 +35,7 @@ namespace classification
 {
 namespace prediction
 {
-namespace interface2
+namespace interface1
 {
 template <typename algorithmFPType, Method method, CpuType cpu>
 BatchContainer<algorithmFPType, method, cpu>::BatchContainer(daal::services::Environment::env * daalEnv) : PredictionContainerIface()
@@ -53,20 +53,17 @@ template <typename algorithmFPType, Method method, CpuType cpu>
 services::Status BatchContainer<algorithmFPType, method, cpu>::compute()
 {
     const classifier::prediction::Input * const input = static_cast<const classifier::prediction::Input *>(_in);
-    classifier::prediction::Result * const result = static_cast<classifier::prediction::Result *>(_res);
-    classifier::Parameter * const parameter = static_cast<classifier::Parameter *>(_par);
+    classifier::prediction::interface1::Result * const result = static_cast<classifier::prediction::interface1::Result *>(_res);
+    classifier::interface1::Parameter * const parameter = static_cast<classifier::interface1::Parameter *>(_par);
 
     const data_management::NumericTableConstPtr a = input->get(classifier::prediction::data);
     const classifier::ModelConstPtr m = input->get(classifier::prediction::model);
     const data_management::NumericTablePtr r = result->get(classifier::prediction::prediction);
-    data_management::NumericTablePtr prob; // Used to prevent shared pointer release
-    data_management::NumericTable * const p = ((parameter->resultsToEvaluate & classifier::computeClassProbabilities) != 0)
-                                              ? (prob = result->get(classifier::prediction::probabilities)).get() : nullptr;
 
     daal::services::Environment::env & env = *_env;
 
     __DAAL_CALL_KERNEL(env, internal::DecisionTreePredictKernel, __DAAL_KERNEL_ARGUMENTS(algorithmFPType, method), \
-                       compute, a.get(), m.get(), r.get(), p, parameter->nClasses);
+                       compute, a.get(), m.get(), r.get(), 0, parameter->nClasses);
 }
 }
 } // namespace prediction
