+++ algorithms/kernel/decision_tree/inner/decision_tree_classification_predict_dense_default_batch_fpt_cpu_v1.cpp
@@ -1,4 +1,4 @@
-/* file: decision_tree_regression_predict_dense_default_batch_fpt_cpu.cpp */
+/* file: decision_tree_classification_predict_dense_default_batch_fpt_cpu_v1.cpp */
 /*******************************************************************************
 * Copyright 2014-2019 Intel Corporation
 *
@@ -21,9 +21,9 @@
 //--
 */
 
-#include "decision_tree_regression_predict_dense_default_batch.h"
-#include "decision_tree_regression_predict_dense_default_batch_impl.i"
-#include "decision_tree_regression_predict_dense_default_batch_container.h"
+#include "decision_tree_classification_predict_dense_default_batch.h"
+#include "decision_tree_classification_predict_dense_default_batch_impl.i"
+#include "decision_tree_classification_predict_dense_default_batch_container_v1.h"
 
 namespace daal
 {
@@ -31,22 +31,16 @@ namespace algorithms
 {
 namespace decision_tree
 {
-namespace regression
+namespace classification
 {
 namespace prediction
 {
-
-namespace interface2
+namespace interface1
 {
 template class BatchContainer<DAAL_FPTYPE, defaultDense, DAAL_CPU>;
-} // namespace interface2
-
-namespace internal
-{
-template class DecisionTreePredictKernel<DAAL_FPTYPE, defaultDense, DAAL_CPU>;
-} // namespace internal
+}
 } // namespace prediction
-} // namespace regression
+} // namespace classification
 } // namespace decision_tree
 } // namespace algorithms
 } // namespace daal
