+++ algorithms/kernel/decision_tree/inner/decision_tree_regression_predict_dense_default_batch_container_v1.h
@@ -1,4 +1,4 @@
-/* file: decision_tree_classification_predict_dense_default_batch_container.h */
+/* file: decision_tree_regression_predict_dense_default_batch_container_v1.h */
 /*******************************************************************************
 * Copyright 2014-2019 Intel Corporation
 *
@@ -22,8 +22,8 @@
 //--
 */
 
-#include "decision_tree_classification_predict.h"
-#include "decision_tree_classification_predict_dense_default_batch.h"
+#include "decision_tree_regression_predict.h"
+#include "decision_tree_regression_predict_dense_default_batch.h"
 
 namespace daal
 {
@@ -31,12 +31,13 @@ namespace algorithms
 {
 namespace decision_tree
 {
-namespace classification
+namespace regression
 {
 namespace prediction
 {
-namespace interface2
+namespace interface1
 {
+
 template <typename algorithmFPType, Method method, CpuType cpu>
 BatchContainer<algorithmFPType, method, cpu>::BatchContainer(daal::services::Environment::env * daalEnv) : PredictionContainerIface()
 {
@@ -52,25 +53,22 @@ BatchContainer<algorithmFPType, method, cpu>::~BatchContainer()
 template <typename algorithmFPType, Method method, CpuType cpu>
 services::Status BatchContainer<algorithmFPType, method, cpu>::compute()
 {
-    const classifier::prediction::Input * const input = static_cast<const classifier::prediction::Input *>(_in);
-    classifier::prediction::Result * const result = static_cast<classifier::prediction::Result *>(_res);
-    classifier::Parameter * const parameter = static_cast<classifier::Parameter *>(_par);
+    const Input * const input = static_cast<const Input *>(_in);
+    Result * const result = static_cast<Result *>(_res);
 
-    const data_management::NumericTableConstPtr a = input->get(classifier::prediction::data);
-    const classifier::ModelConstPtr m = input->get(classifier::prediction::model);
-    const data_management::NumericTablePtr r = result->get(classifier::prediction::prediction);
-    data_management::NumericTablePtr prob; // Used to prevent shared pointer release
-    data_management::NumericTable * const p = ((parameter->resultsToEvaluate & classifier::computeClassProbabilities) != 0)
-                                              ? (prob = result->get(classifier::prediction::probabilities)).get() : nullptr;
+    const data_management::NumericTableConstPtr a = input->get(data);
+    const ModelConstPtr m = input->get(model);
+    const data_management::NumericTablePtr r = result->get(prediction);
 
+    const daal::algorithms::Parameter * const par = _par;
     daal::services::Environment::env & env = *_env;
 
     __DAAL_CALL_KERNEL(env, internal::DecisionTreePredictKernel, __DAAL_KERNEL_ARGUMENTS(algorithmFPType, method), \
-                       compute, a.get(), m.get(), r.get(), p, parameter->nClasses);
+                       compute, a.get(), m.get(), r.get(), par);
 }
 }
 } // namespace prediction
-} // namespace classification
+} // namespace regression
 } // namespace decision_tree
 } // namespace algorithms
 } // namespace daal
