+++ algorithms/kernel/decision_tree/inner/decision_tree_regression_predict_dense_default_batch_fpt_cpu_v1.cpp
@@ -1,4 +1,4 @@
-/* file: decision_tree_regression_predict_dense_default_batch_fpt_cpu.cpp */
+/* file: decision_tree_regression_predict_dense_default_batch_fpt_cpu_v1.cpp */
 /*******************************************************************************
 * Copyright 2014-2019 Intel Corporation
 *
@@ -23,7 +23,7 @@
 
 #include "decision_tree_regression_predict_dense_default_batch.h"
 #include "decision_tree_regression_predict_dense_default_batch_impl.i"
-#include "decision_tree_regression_predict_dense_default_batch_container.h"
+#include "decision_tree_regression_predict_dense_default_batch_container_v1.h"
 
 namespace daal
 {
@@ -35,16 +35,10 @@ namespace regression
 {
 namespace prediction
 {
-
-namespace interface2
+namespace interface1
 {
 template class BatchContainer<DAAL_FPTYPE, defaultDense, DAAL_CPU>;
-} // namespace interface2
-
-namespace internal
-{
-template class DecisionTreePredictKernel<DAAL_FPTYPE, defaultDense, DAAL_CPU>;
-} // namespace internal
+} // namespace interface1
 } // namespace prediction
 } // namespace regression
 } // namespace decision_tree
