+++ algorithms/kernel/decision_tree/inner/decision_tree_regression_train_container_v1.h
@@ -1,4 +1,4 @@
-/* file: decision_tree_classification_train_container.h */
+/* file: decision_tree_regression_train_container_v1.h */
 /*******************************************************************************
 * Copyright 2014-2019 Intel Corporation
 *
@@ -21,15 +21,15 @@
 //--
 */
 
-#ifndef __DECISION_TREE_CLASSIFICATION_TRAIN_CONTAINER_H__
-#define __DECISION_TREE_CLASSIFICATION_TRAIN_CONTAINER_H__
+#ifndef __DECISION_TREE_REGRESSION_TRAIN_CONTAINER_H__
+#define __DECISION_TREE_REGRESSION_TRAIN_CONTAINER_H__
 
 #include "kernel.h"
 #include "data_management/data/numeric_table.h"
 #include "services/daal_shared_ptr.h"
-#include "decision_tree_classification_training_batch.h"
-#include "decision_tree_classification_train_kernel.h"
-#include "decision_tree_classification_model_impl.h"
+#include "decision_tree_regression_training_batch.h"
+#include "decision_tree_regression_train_kernel.h"
+#include "decision_tree_regression_model_impl.h"
 
 namespace daal
 {
@@ -37,11 +37,11 @@ namespace algorithms
 {
 namespace decision_tree
 {
-namespace classification
+namespace regression
 {
 namespace training
 {
-namespace interface2
+namespace interface1
 {
 using namespace daal::data_management;
 
@@ -51,7 +51,7 @@ using namespace daal::data_management;
 template <typename algorithmFPType, training::Method method, CpuType cpu>
 BatchContainer<algorithmFPType, method, cpu>::BatchContainer(daal::services::Environment::env *daalEnv)
 {
-    __DAAL_INITIALIZE_KERNELS(internal::DecisionTreeTrainBatchKernel, decision_tree::classification::Parameter, algorithmFPType, method);
+    __DAAL_INITIALIZE_KERNELS(internal::DecisionTreeTrainBatchKernel, algorithmFPType, method);
 }
 
 template <typename algorithmFPType, training::Method method, CpuType cpu>
@@ -66,26 +66,25 @@ BatchContainer<algorithmFPType, method, cpu>::~BatchContainer()
 template <typename algorithmFPType, training::Method method, CpuType cpu>
 services::Status BatchContainer<algorithmFPType, method, cpu>::compute()
 {
-    const decision_tree::classification::training::Input * const input = static_cast<decision_tree::classification::training::Input *>(_in);
+    const decision_tree::regression::training::Input * const input = static_cast<decision_tree::regression::training::Input *>(_in);
     Result * const result = static_cast<Result *>(_res);
 
-    const NumericTableConstPtr x = input->get(classifier::training::data);
-    const NumericTableConstPtr y = input->get(classifier::training::labels);
-    const NumericTableConstPtr w = input->get(classifier::training::weights);
+    const NumericTableConstPtr x = input->get(data);
+    const NumericTableConstPtr y = input->get(dependentVariables);
     const NumericTableConstPtr px = input->get(dataForPruning);
-    const NumericTableConstPtr py = input->get(labelsForPruning);
+    const NumericTableConstPtr py = input->get(dependentVariablesForPruning);
 
-    const ModelPtr r = result->get(classifier::training::model);
+    const ModelPtr r = result->get(model);
 
-    const Parameter * const par = static_cast<Parameter*>(_par);
+    const daal::algorithms::Parameter * const par = _par;
     daal::services::Environment::env & env = *_env;
 
-    __DAAL_CALL_KERNEL(env, internal::DecisionTreeTrainBatchKernel, __DAAL_KERNEL_ARGUMENTS(algorithmFPType, decision_tree::classification::Parameter,method),    \
-                       compute, x.get(), y.get(), w.get(), px.get(), py.get(), r.get(), par);
+    __DAAL_CALL_KERNEL(env, internal::DecisionTreeTrainBatchKernel, __DAAL_KERNEL_ARGUMENTS(algorithmFPType, method),    \
+                       compute, x.get(), y.get(), 0, px.get(), py.get(), r.get(), par);
 }
 }
 } // namespace training
-} // namespace classification
+} // namespace regression
 } // namespace decision_tree
 } // namespace algorithms
 } // namespace daal
