+++ makefile.lst
@@ -47,7 +47,7 @@ math += math/abs math/logistic math/relu math/smoothrelu math/softmax math/tanh
 optimization_solver += optimization_solver/adagrad optimization_solver/adagrad/inner optimization_solver/lbfgs optimization_solver/lbfgs/inner optimization_solver/sgd optimization_solver/sgd/inner optimization_solver/saga optimization_solver/saga/inner optimization_solver/inner optimization_solver/coordinate_descent objective_function engines distributions
 coordinate_descent += optimization_solver/coordinate_descent objective_function engines distributions
 objective_function += objective_function/inner objective_function/cross_entropy_loss objective_function/cross_entropy_loss/inner objective_function/logistic_loss objective_function/logistic_loss/inner objective_function/mse objective_function/mse/inner
-decision_tree += regression classifier
+decision_tree += decision_tree/inner regression classifier
 dtrees/gbt += dtrees dtrees/gbt/classification dtrees/gbt/regression engines classifier regression objective_function
 dtrees/forest += dtrees dtrees/regression dtrees/forest/classification dtrees/forest/regression engines classifier regression distributions
 linear_regression += linear_model regression
